+++ src/actions/orgUnits.js
@@ -13,39 +13,6 @@ export const setOrgUnitTree = (rootModel) => ({
     payload: rootModel,
 })
 
-// Load all organisation unit groups sets
-export const loadOrgUnitLevels = () => ({
-    type: types.ORGANISATION_UNIT_LEVELS_LOAD,
-})
-
-// Set all organisation unit groups sets
-export const setOrgUnitLevels = (data) => ({
-    type: types.ORGANISATION_UNIT_LEVELS_SET,
-    payload: data,
-})
-
-// Load all organisation unit groups sets
-export const loadOrgUnitGroups = () => ({
-    type: types.ORGANISATION_UNIT_GROUPS_LOAD,
-})
-
-// Set all organisation unit groups sets
-export const setOrgUnitGroups = (data) => ({
-    type: types.ORGANISATION_UNIT_GROUPS_SET,
-    payload: data,
-})
-
-// Load all organisation unit groups sets
-export const loadOrgUnitGroupSets = () => ({
-    type: types.ORGANISATION_UNIT_GROUP_SETS_LOAD,
-})
-
-// Set all organisation unit groups sets
-export const setOrgUnitGroupSets = (data) => ({
-    type: types.ORGANISATION_UNIT_GROUP_SETS_SET,
-    payload: data,
-})
-
 export const setOrgUnitProfile = (id) => ({
     type: types.ORGANISATION_UNIT_PROFILE_SET,
     payload: id,