+++ src/components/BaseUrlShim.js
@@ -1,0 +1,7 @@
+import { useConfig } from '@dhis2/app-runtime'
+
+export const BaseUrlShim = ({ children }) => {
+    const { baseUrl, apiVersion } = useConfig()
+
+    return children({ baseUrl: `${baseUrl}/api/${apiVersion}` })
+}