+++ src/components/classification/LegendSetSelect.js
@@ -1,54 +1,46 @@
+import { useDataQuery } from '@dhis2/app-runtime'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from 'prop-types'
-import React, { Component } from 'react'
-import { connect } from 'react-redux'
+import React from 'react'
+import { useSelector, useDispatch } from 'react-redux'
 import { setLegendSet } from '../../actions/layerEdit.js'
-import { loadLegendSets } from '../../actions/legendSets.js'
 import { SelectField } from '../core/index.js'
+
+// Load all legend sets
+const LEGEND_SETS_QUERY = {
+    sets: {
+        resource: 'legendSets',
+        params: {
+            fields: ['id', 'displayName~rename(name)'],
+            paging: false,
+        },
+    },
+}
 
 const style = {
     width: '100%',
 }
 
-class LegendSetSelect extends Component {
-    static propTypes = {
-        loadLegendSets: PropTypes.func.isRequired,
-        setLegendSet: PropTypes.func.isRequired,
-        legendSet: PropTypes.object,
-        legendSetError: PropTypes.string,
-        legendSets: PropTypes.array,
-    }
+const LegendSetSelect = ({ legendSetError }) => {
+    const legendSet = useSelector((state) => state.layerEdit.legendSet)
+    const dispatch = useDispatch()
+    const { loading, error, data } = useDataQuery(LEGEND_SETS_QUERY)
 
-    componentDidMount() {
-        const { legendSets, loadLegendSets } = this.props
-
-        if (!legendSets) {
-            loadLegendSets()
-        }
-    }
-
-    render() {
-        const { legendSet, legendSets, legendSetError, setLegendSet } =
-            this.props
-
-        return (
-            <SelectField
-                label={i18n.t('Legend set')}
-                loading={legendSets ? false : true}
-                items={legendSets}
-                value={legendSet ? legendSet.id : null}
-                errorText={legendSetError}
-                onChange={setLegendSet}
-                style={style}
-            />
-        )
-    }
+    return (
+        <SelectField
+            label={i18n.t('Legend set')}
+            loading={loading}
+            items={data?.sets.legendSets}
+            value={legendSet ? legendSet.id : null}
+            errorText={error?.message || legendSetError}
+            onChange={(legendSet) => dispatch(setLegendSet(legendSet))}
+            style={style}
+        />
+    )
 }
 
-export default connect(
-    (state) => ({
-        legendSet: state.layerEdit.legendSet,
-        legendSets: state.legendSets,
-    }),
-    { loadLegendSets, setLegendSet }
-)(LegendSetSelect)
+LegendSetSelect.propTypes = {
+    legendSetError: PropTypes.string,
+}
+
+export default LegendSetSelect